refactor: http error와 error 핸들러 수정
