+++ server/src/app.ts
@@ -6,7 +6,7 @@ import cookieParser from "cookie-parser";
 import { Connection, createConnection } from "typeorm";
 import postRouter from "./routes/post-router";
 import channelRouter from "./routes/chnnel-router";
-import errorHandler from "./middlewares/errorHandler";
+import errorHandler from "./middleware/errorHandler";
 
 export default class App {
   private static app: Express;
@@ -24,6 +24,7 @@ export default class App {
   private static initializeExpress() {
     this.app = express();
     this.app.set("port", process.env.PORT || 3000);
+    this.app.set("env", process.env.NODE_ENV);
     this.app.use(morgan("dev"));
     this.app.use(express.json());
     this.app.use(express.urlencoded({ extended: false }));
