--- server/src/exception/HttpException.ts
@@ -1,9 +1,0 @@
-export default class HttpException extends Error {
-    public status: number;
-    public message: string;
-    constructor(status: number, message: string) {
-        super(message);
-        this.status = status;
-        this.message = message;
-    }
-}
