+++ server/src/util/HttpException.ts
@@ -1,0 +1,8 @@
+export default class HttpException extends Error {
+    public status?: number;
+    
+    constructor(message: string, status?: number,) {
+        super(message);
+        this.status = status;
+    }
+}
